fix linkall folder creation

+++ linkall.sh
@@ -5,13 +5,17 @@ echo ''
 
 HERE=$(dirname $0)
 echo '>>> sym-linking config files' $HERE 'to' $HOME
-rm -r $HOME/{.vimrc,.zshrc,.tmux.conf,.bash_profile,.taskrc,.tmuxinator}
-ln -sv $HERE/{.vimrc,.zshrc,.tmux.conf,.bash_profile,.taskrc,.tmuxinator} $HOME
+rm -r $HOME/{.vimrc,.zshrc,.tmux.conf,.bash_profile,.taskrc,.tmuxinator,.gitconfig}
+ln -sv $HERE/{.vimrc,.zshrc,.tmux.conf,.bash_profile,.taskrc,.tmuxinator,.gitconfig} $HOME
 echo ''
 
 # Link vundle 
 echo '>>> sym-linking vundle'
-rm -rf $HOME/.vim/bundle/vundle
+if [ -d $HOME/.vim/bundle/vundle ]; then
+  rm -rf $HOME/.vim/bundle/vundle
+else
+  mkdir -pv $HOME/.vim/bundle/vundle
+fi
 ln -sv $HERE/vundle $HOME/.vim/bundle/vundle
 echo ''
 
@@ -24,12 +28,15 @@ echo ''
 # links not in home directory
 # Elinks conf
 echo '>>> sym-linking elinks.conf'
-rm $HOME/.elinks/elinks.conf
+if [ -d $HOME/.elinks/ ]; then
+  rm -v $HOME/.elinks/elinks.conf
+else
+  mkdir -pv $HOME/.elinks/elinks.conf
+fi
 ln -sv $HERE/elinks.conf $HOME/.elinks/elinks.conf
 echo ''
 
 # weechat scripts and conf (ignore the log files), 
-# pride: rewrite this later to be less silly
 echo '>>> sym-linking weechat folder (maintaining log files)'
 mv $HOME/.weechat/logs $HOME/temp-logs/
 mv $HOME/.weechat/weechat.log $HOME/temp-weechat.log
@@ -51,20 +58,23 @@ echo ''
 
 # matlab startup file
 echo '>>> sym-linking matlab startup'
-rm $HOME/Documents/MATLAB/startup.m
+if [ -d $HOME/Documents/MATLAB ]; then
+  rm -v $HOME/Documents/MATLAB/startup.m
+else
+  mkdir -pv $HOME/Documents/MATLAB/startup.m
+fi
 ln -sv $HERE/startup.m $HOME/Documents/MATLAB/startup.m
 echo ''
 
 # vifm (greatest file browser ever made) link, don't link
 # or overwrite history, Trash, help files etc.,
 echo '>>> sym-linking vifm files'
-rm -rf $HOME/.vifm/colors
-ln -sv $HERE/.vifm/colors $HOME/.vifm/colors
-rm -rf $HOME/.vifm/vifmrc
+if [ -d $HOME/.vifm ]; then
+    rm -rf $HOME/.vifm/colors
+    rm -rf $HOME/.vifm/vifmrc
+else
+    mkdir -pv $HOME/.vifm/
+fi
 ln -sv $HERE/.vifm/vifmrc $HOME/.vifm/vifmrc
+ln -sv $HERE/.vifm/colors $HOME/.vifm/colors
 echo ''
-
-# update my vim bundles
-#echo '>>> Updating vim bundles'
-#vim +BundleInstall! +BundleClean
-#echo ''
